+++ internal/migrate/is_staged.go
@@ -30,6 +30,7 @@ import (
 
 	"github.com/onflow/flow-cli/internal/command"
 	"github.com/onflow/flow-cli/internal/scripts"
+	"github.com/onflow/flow-cli/internal/util"
 )
 
 var isStagedflags struct{}
@@ -52,13 +53,13 @@ func isStaged(
 	flow flowkit.Services,
 	state *flowkit.State,
 ) (command.Result, error) {
-	err := checkNetwork(flow.Network())
+	err := util.CheckNetwork(flow.Network())
 	if err != nil {
 		return nil, err
 	}
 
 	contractName := args[0]
-	addr, err := getAddressByContractName(state, contractName, flow.Network())
+	addr, err := util.GetAddressByContractName(state, contractName, flow.Network())
 	if err != nil {
 		return nil, fmt.Errorf("error getting address by contract name: %w", err)
 	}
