+++ internal/migrate/state.go
@@ -43,6 +43,7 @@ import (
 	"github.com/onflow/flow-go-sdk"
 
 	"github.com/onflow/flow-cli/internal/command"
+	"github.com/onflow/flow-cli/internal/util"
 )
 
 var stateFlags struct {
@@ -147,7 +148,7 @@ func resolveStagedContracts(state *flowkit.State, contractNames []string) ([]mig
 
 		// If contract is not aliased, try to get address by deployment account
 		if address == flow.EmptyAddress {
-			address, err = getAddressByContractName(state, contractName, network)
+			address, err = util.GetAddressByContractName(state, contractName, network)
 			if err != nil {
 				return nil, fmt.Errorf("failed to get address by contract name: %w", err)
 			}
